requirements.txt
numpy
pandas
requests
streamlit
streamlit-webrtc
watchdog
# Main dependencies for inference models only
inference-sdk
opencv-python
python-dotenv
ultralytics